Reload virus scan engine upon database update
If freshclam has changed the virus pattern database a new virus
scanning enging is loaded and the cache is cleared.

For testing call skyldav with -m 1.
Use
sudo touch -m /var/lib/clamav/main.cvd
to simulated an update of the pattern database.

If a new device is mounted the cache is cleared as it may
have been modified while being offline (e.g. USB stick).

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>